feat: migrate database from PostgreSQL to SQLite
Replace PostgreSQL/Kysely with SQLite for a self-contained, zero-dependency database.
Removes psql utilities, PostgreSQL error codes, and hostname formatting. Adds SQLite
error codes, unique violation helper, and ISO date conversion. Simplifies connection
UI by removing server/port/credentials fields. Updates build config, migrations, and
all query files for SQLite compatibility.

+++ src/node/database/migrations/v1/create-checksum-tags-table.ts
@@ -7,7 +7,7 @@ const createChecksumTagsTable: Migration = {
       .createTable('checksum_tags')
       .ifNotExists()
       .addColumn('checksum', 'varchar', (col) => col.notNull())
-      .addColumn('tag_id', 'int8', (col) => col.notNull())
+      .addColumn('tag_id', 'integer', (col) => col.notNull())
       .addUniqueConstraint('checksum_tags_checksum_tag_id_unique', ['checksum', 'tag_id'])
       .execute();
   },
